complete create and edit routes

+++ labconnect/main/opportunity_routes.py
@@ -496,25 +496,6 @@ def getOpportunityCards():
 #     abort(500)
 
 
-# @main_blueprint.route("/getOpportunityByProfessor/<string:rcs_id>", methods=["GET"])
-# def getOpportunityByProfessor(rcs_id: str):
-#     if request.method == "GET":
-#         # query database for opportunity
-#         query = db.session.execute(
-#             db.select(Opportunities, Leads)
-#             .where(Leads.lab_manager_id == rcs_id)
-#             .join(Opportunities, Leads.opportunity_id == Opportunities.id)
-#         )
-
-#         data = query.all()
-#         print(data)
-
-#         # return data in the below format if opportunity is found
-#         return {"data": [opportunity[0].to_dict() for opportunity in data]}
-
-#     abort(500)
-
-
 @main_blueprint.get("/staff/opportunities/<string:rcs_id>")
 def getLabManagerOpportunityCards(rcs_id: str):
 
@@ -633,53 +614,41 @@ def createOpportunity():
     if lenum is None:
         lenum = LocationEnum.TBD
 
-    newOpportunity = Opportunities(
-        name=request_data["title"],
-        description=request_data["description"],
-        recommended_experience=request_data["recommended_experience"],
-        pay=pay,
-        one_credit=one,
-        two_credits=two,
-        three_credits=three,
-        four_credits=four,
-        semester=SemesterEnum.FALL,
-        year=datetime.now().year,
-        application_due=datetime.strptime(request_data["application_due"], "%Y-%m-%d"),
-        active=True,
-        location=lenum,
-        last_updated=datetime.now(),
+    newOpportunity = Opportunities()
+    newOpportunity.name = request_data["title"]
+    newOpportunity.description = request_data["description"]
+    newOpportunity.recommended_experience = request_data["recommended_experience"]
+    newOpportunity.pay = pay
+    newOpportunity.one_credit = one
+    newOpportunity.two_credits = two
+    newOpportunity.three_credits = three
+    newOpportunity.four_credits = four
+    newOpportunity.semester = SemesterEnum.FALL
+    newOpportunity.year = datetime.now().year
+    newOpportunity.application_due = datetime.strptime(
+        request_data["application_due"], "%Y-%m-%d"
     )
+    newOpportunity.active = True
+    newOpportunity.location = lenum
+    newOpportunity.last_updated = datetime.now()
     db.session.add(newOpportunity)
     db.session.commit()
 
-    newLead = Leads(
-        lab_manager_id=author.lab_manager_id, opportunity_id=newOpportunity.id
-    )
+    newLead = Leads()
+    newLead.lab_manager_id = author.lab_manager_id
+    newLead.opportunity_id = newOpportunity.id
 
     db.session.add(newLead)
-    db.session.commit()
-
-    # for course in newPostData["courses"]:
-    #     newCourse = RecommendsCourses(
-    #         opportunity_id=newOpportunity.id, course_code=course
-    #     )
-
-    #     db.session.add(newCourse)
-    #     db.session.commit()
-
-    # for major in newPostData["majors"]:
-    #     newMajor = RecommendsMajors(opportunity_id=newOpportunity.id, major_code=major)
-    #     db.session.add(newMajor)
-    #     db.session.commit()
 
     for year in request_data["years"]:
         if year.isdigit():
             recommended_year = int(year)
-            newYear = RecommendsClassYears(
-                opportunity_id=newOpportunity.id, class_year=recommended_year
-            )
+            newYear = RecommendsClassYears()
+            newYear.opportunity_id = newOpportunity.id
+            newYear.class_year = recommended_year
             db.session.add(newYear)
-            db.session.commit()
+
+    db.session.commit()
 
     return {"data": "Opportunity Created", "id": newOpportunity.id}, 200
 
@@ -730,6 +699,8 @@ def editOpportunity_get(opportunity_id):
         if credit
     ]
 
+    years = [str(year.class_year) for year in years_data] if years_data else []
+
     # Format opportunity data as JSON
     opportunity_data = {
         "id": opportunity.id,
@@ -737,8 +708,8 @@ def editOpportunity_get(opportunity_id):
         "application_due": opportunity.application_due.strftime("%Y-%m-%d"),
         "type": (
             "Any"
-            if len(credits) > 0 and opportunity.pay > 0
-            else "For Pay" if opportunity.pay > 0 else "For Credit"
+            if len(credits) > 0 and opportunity.pay and opportunity.pay > 0
+            else "For Pay" if opportunity.pay and opportunity.pay > 0 else "For Credit"
         ),
         "hourlyPay": str(opportunity.pay),
         "credits": credits,
@@ -751,7 +722,7 @@ def editOpportunity_get(opportunity_id):
         # "last_updated": opportunity.last_updated.strftime("%Y-%m-%d %H:%M:%S"),
         # "courses": [course.course_code for course in courses_data],
         # "majors": [major.major_code for major in majors_data],
-        "years": [str(year.class_year) for year in years_data],
+        "years": years,
     }
 
     return opportunity_data
@@ -761,102 +732,100 @@ def editOpportunity_get(opportunity_id):
 @jwt_required()
 def editOpportunity(opportunity_id):
 
-    if not request.data:
+    user_id = get_jwt_identity()
+    if not request.data or not user_id:
         abort(400)
 
-    data = request.get_json()
+    request_data = request.get_json()
 
-    if not data:
+    if not request_data:
         abort(400)
 
     # Check if the opportunity and author exist
     opportunity = db.session.execute(
         db.select(Opportunities).where(Opportunities.id == opportunity_id)
-    ).first()
-
-    if not opportunity:
-        return {"error": "Opportunity not found"}, 404
-
-    opportunity = opportunity[0]
-    data = data[0]
-
-    # TODO: Add check to see if person has permission to edit opportunity
-    user_id = get_jwt_identity()
-
-    lab_manager = db.session.execute(
-        db.select(LabManager)
-        .join(User, User.lab_manager_id == LabManager.id)
-        .where(User.email == user_id)
     ).scalar_one_or_none()
 
-    if lab_manager:
-        lead = db.session.execute(
-            db.select(Leads).where(
-                Leads.lab_manager_id == lab_manager.id,
-                Leads.opportunity_id == opportunity.id,
-            )
-        ).scalar_one_or_none()
-        if not lead:
-            return {"error": "Don't have permission to edit!"}, 401
-    else:
-        return {"error": "Don't have permission to edit!"}, 401
+    if opportunity is None:
+        abort(400)
+
+    author = db.session.execute(
+        db.select(User).where(User.email == user_id[0])
+    ).scalar_one_or_none()
+
+    if author is None or author.lab_manager_id is None:
+        abort(400)
+
+    leads = db.session.execute(
+        db.select(Leads)
+        .where(Leads.opportunity_id == opportunity_id)
+        .where(Leads.lab_manager_id == author.lab_manager_id)
+    ).scalar_one_or_none()
+
+    if leads is None:
+        abort(400)
+
+    try:
+        pay = int(request_data["hourlyPay"])
+    except:
+        pay = None
+
+    one = True if "1" in request_data["credits"] else False
+    two = True if "2" in request_data["credits"] else False
+    three = True if "3" in request_data["credits"] else False
+    four = True if "4" in request_data["credits"] else False
+
+    lenum = convert_to_enum(request_data["location"])
+
+    if lenum is None:
+        lenum = LocationEnum.TBD
 
     # Update fields for opportunity based on the input data
-    opportunity.name = data["name"]
-    opportunity.description = data["description"]
-    opportunity.recommended_experience = data["recommended_experience"]
-    opportunity.pay = data["pay"]
-    opportunity.one_credit = data["one_credit"]
-    opportunity.two_credits = data["two_credits"]
-    opportunity.three_credits = data["three_credits"]
-    opportunity.four_credits = data["four_credits"]
-    opportunity.semester = (
-        SemesterEnum[(data["semester"]).upper()]
-        if "semester" in data
-        else opportunity.semester
+    opportunity.name = request_data["title"]
+    opportunity.description = request_data["description"]
+    opportunity.recommended_experience = request_data["recommended_experience"]
+    opportunity.pay = pay
+    opportunity.one_credit = one
+    opportunity.two_credits = two
+    opportunity.three_credits = three
+    opportunity.four_credits = four
+    opportunity.application_due = datetime.strptime(
+        request_data["application_due"], "%Y-%m-%d"
     )
-    opportunity.year = data["year"]
-    opportunity.application_due = (
-        datetime.datetime.strptime(data["application_due"], "%Y-%m-%d")
-        if "application_due" in data
-        else opportunity.application_due
-    )
-    opportunity.active = data["active"]
-    opportunity.location = (
-        convert_to_enum(data["location"])
-        if "location" in data
-        else opportunity.location
-    )
-    opportunity.last_updated = datetime.datetime.now()
+    # opportunity.active = data["active"]
+    opportunity.location = lenum
+    opportunity.last_updated = datetime.now()
 
-    # Update related tables for courses, majors, and years
-    # Clear current recommendations
-    db.session.query(RecommendsCourses).filter_by(
-        opportunity_id=opportunity_id
-    ).delete()
-    db.session.query(RecommendsMajors).filter_by(opportunity_id=opportunity_id).delete()
-    db.session.query(RecommendsClassYears).filter_by(
-        opportunity_id=opportunity_id
-    ).delete()
+    existing_years = {
+        str(year.class_year)
+        for year in db.session.execute(
+            db.select(RecommendsClassYears).where(
+                RecommendsClassYears.opportunity_id == opportunity_id
+            )
+        ).scalars()
+    }
+    new_years = set(request_data["years"])
 
-    # Re-add new recommendations
-    for course in data["courses"]:
-        newCourse = RecommendsCourses(opportunity_id=opportunity.id, course_code=course)
-
-        db.session.add(newCourse)
-        db.session.commit()
-
-    for major in data["majors"]:
-        newMajor = RecommendsMajors(opportunity_id=opportunity.id, major_code=major)
-        db.session.add(newMajor)
-        db.session.commit()
-
-    for year in data["years"]:
-        newYear = RecommendsClassYears(opportunity_id=opportunity.id, class_year=year)
+    # Years to add
+    years_to_add = new_years - existing_years
+    for year in years_to_add:
+        newYear = RecommendsClassYears()
+        newYear.opportunity_id = opportunity.id
+        newYear.class_year = int(year)
         db.session.add(newYear)
-        db.session.commit()
 
-    # Commit all changes to the database
+    # Years to remove
+    years_to_remove = existing_years - new_years
+    if years_to_remove:
+        db.session.execute(
+            db.select(RecommendsClassYears)
+            .where(
+                RecommendsClassYears.opportunity_id == opportunity_id,
+                RecommendsClassYears.class_year.in_(years_to_remove),
+            )
+            .delete(synchronize_session=False)
+        )
+
     db.session.commit()
 
     return {"data": "Opportunity Updated"}, 200
@@ -873,35 +842,24 @@ def deleteOpportunity(opportunity_id):
     # TODO: Add check to see if user has permission to delete opportunity
     user_id = get_jwt_identity()
 
-    lab_manager = db.session.execute(
-        db.select(LabManager)
-        .join(User, User.lab_manager_id == LabManager.id)
-        .where(User.email == user_id)
+    user = db.session.execute(
+        db.select(User).where(User.email == user_id)
     ).scalar_one_or_none()
 
-    if lab_manager:
-        lead = db.session.execute(
-            db.select(Leads).where(
-                Leads.lab_manager_id == lab_manager.id,
-                Leads.opportunity_id == opportunity.id,
-            )
-        ).scalar_one_or_none()
-        if not lead:
-            return {"error": "Don't have permission to delete!"}, 401
-    else:
+    if not user or not user.lab_manager_id:
         return {"error": "Don't have permission to delete!"}, 401
 
-    # Delete related records in other tables (e.g., Leads, RecommendsCourses, RecommendsMajors, RecommendsClassYears)
-    db.session.query(Leads).filter_by(opportunity_id=opportunity_id).delete()
-    db.session.query(RecommendsCourses).filter_by(
-        opportunity_id=opportunity_id
-    ).delete()
-    db.session.query(RecommendsMajors).filter_by(opportunity_id=opportunity_id).delete()
-    db.session.query(RecommendsClassYears).filter_by(
-        opportunity_id=opportunity_id
-    ).delete()
+    leads = db.session.execute(
+        db.select(Leads)
+        .where(Leads.opportunity_id == opportunity_id)
+        .where(Leads.lab_manager_id == user.lab_manager_id)
+    ).scalar_one_or_none()
 
-    # Delete the opportunity itself
+    if not leads:
+        abort(400)
+
+    # Delete the opportunity
+    # cascading delete will handle all other tables
     db.session.delete(opportunity)
     db.session.commit()
